move views to component folder
