+++ client/app/help/components/HearingsHelp.jsx
